backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.8.2
pydantic[email]

# AI chat — Groq is 100% free, no credit card needed
# sign up at console.groq.com
groq==0.9.0

# Email — Resend free tier: 3,000 emails/month, no credit card
# sign up at resend.com
resend==2.2.0

# Security
slowapi==0.1.9
bleach==6.1.0

# HTTP
httpx==0.27.0
